app: simplify sculpt controls
