+++ packages/app/src/style.css
@@ -424,85 +424,3 @@ body,
   text-transform: uppercase;
   letter-spacing: 0.05em;
 }
-
-.tools-controls {
-  position: absolute;
-  top: 24px;
-  left: calc(16px + 52px + 12px);
-  display: flex;
-  gap: 12px;
-  align-items: flex-start;
-  z-index: 6;
-  transition: opacity 0.2s ease, transform 0.2s ease;
-}
-
-.tools-controls-hidden {
-  opacity: 0;
-  transform: translateX(-8px);
-  pointer-events: none;
-}
-
-.tools-control {
-  display: flex;
-  flex-direction: column;
-  gap: 0;
-  font-size: 0.75rem;
-  color: rgba(255, 255, 255, 0.8);
-  text-transform: uppercase;
-  letter-spacing: 0.06em;
-}
-
-.tools-control-track {
-  position: relative;
-  width: 180px;
-  height: 36px;
-}
-
-.tools-control-label {
-  text-align: center;
-  color: rgba(255, 255, 255, 0.85);
-}
-
-.tools-control-track input[type='range'] {
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  cursor: pointer;
-  background: transparent;
-  -webkit-appearance: none;
-  appearance: none;
-}
-
-.tools-control-track input[type='range']::-webkit-slider-runnable-track {
-  height: 4px;
-  border-radius: 999px;
-  background: rgba(255, 255, 255, 0.25);
-}
-
-.tools-control-track input[type='range']::-webkit-slider-thumb {
-  -webkit-appearance: none;
-  appearance: none;
-  width: 14px;
-  height: 14px;
-  border-radius: 50%;
-  background: #8fd9ff;
-  border: 2px solid #05070b;
-  margin-top: -5px;
-}
-
-.tools-control-track input[type='range']::-moz-range-track {
-  height: 4px;
-  border-radius: 999px;
-  background: rgba(255, 255, 255, 0.25);
-}
-
-.tools-control-track input[type='range']::-moz-range-thumb {
-  width: 14px;
-  height: 14px;
-  border-radius: 50%;
-  background: #8fd9ff;
-  border: 2px solid #05070b;
-  transform: translateY(-2px);
-}
